movement out of constants, jumping still not working quite right

+++ src/MovementComponent.h
@@ -15,4 +15,4 @@ struct MovementComponent : public Component {
 	MovementValues movementValues[Constants::MaxEntities_]; // Saved by ID rather than index
 };
 
-void MovementComponent_Add(MovementComponent* movementComponent, uint32 eid, float xVelocity, float yVelocity, int xAccel, int yAccel);
+void MovementComponent_Add(MovementComponent* movementComponent, uint32 eid, float xVelocity, float yVelocity, float xAccel, float yAccel);
